+++ Diary/UI/TaskDetailScreen/TaskDetailScreenAssembly.swift
@@ -6,15 +6,13 @@ protocol TaskDetailScreenAssemblyProtocol {
 
 final class TaskDetailScreenAssembly: TaskDetailScreenAssemblyProtocol {
     static func assemble(taskData: TasksItem) -> UIViewController {
-        let router = TaskDetailScreenRouter()
-        let presenter = TaskDetailScreenPresenter(router: router, taskData: taskData)
+        let presenter = TaskDetailScreenPresenter(taskData: taskData)
         let tableAdapter = TaskDetailScreenTableAdapter()
         let view = TaskDetailScreenViewController(
             presenter: presenter,
             tableAdapter: tableAdapter
         )
         presenter.view = view
-        tableAdapter.presenter = presenter
         
         return view
     }
