--- Diary/UI/TaskDetailScreen/TaskDetailScreenRouter.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-final class TaskDetailScreenRouter {
-    
-}
