+++ Diary/UI/TaskDetailScreen/View/TableViewCell/TaskDetailTableViewCell.swift
@@ -7,9 +7,9 @@ struct TaskDetailTableViewCellData {
 }
 
 final class TaskDetailTableViewCell: UITableViewCell {
-    @IBOutlet weak var timeLabel: UILabel!
-    @IBOutlet weak var titleLabel: UILabel!
-    @IBOutlet weak var descriptionLabel: UILabel!
+    @IBOutlet private weak var timeLabel: UILabel!
+    @IBOutlet private weak var titleLabel: UILabel!
+    @IBOutlet private weak var descriptionLabel: UILabel!
     
     static let identifier = "TaskDetailTableViewCell"
