+++ Diary/UI/TaskDetailScreen/View/TaskDetailScreenViewController.swift
@@ -5,7 +5,6 @@ final class TaskDetailScreenViewController: UIViewController {
     @IBOutlet private weak var tableView: UITableView!
     private let presenter: TaskDetailScreenPresenterProtocol
     private let tableAdapter: TaskDetailScreenTableAdapter
-    private var currentTask: TaskDetailTableViewCellData?
     
     init(presenter: TaskDetailScreenPresenterProtocol, tableAdapter: TaskDetailScreenTableAdapter) {
         self.presenter = presenter
@@ -27,37 +26,10 @@ final class TaskDetailScreenViewController: UIViewController {
     
     private func prepareTableView() {
         tableView.register(TaskDetailTableViewCell.nib(), forCellReuseIdentifier: TaskDetailTableViewCell.identifier)
-        self.tableView.rowHeight = UITableView.automaticDimension
-        self.tableView.estimatedRowHeight = 50
-    }
-}
-
-extension TaskDetailScreenViewController: UITableViewDelegate {
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        tableView.deselectRow(at: indexPath, animated: true)
-    }
-}
-
-extension TaskDetailScreenViewController: UITableViewDataSource {
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: TaskDetailTableViewCell.identifier, for: indexPath)
-        if let reusableCell = cell as? TaskDetailTableViewCell {
-            guard let currentTask else { return UITableViewCell() }
-            reusableCell.configureCellWith(
-                data: .init(
-                    title: currentTask.title,
-                    date: currentTask.date,
-                    description: currentTask.description
-                )
-            )
-            reusableCell.contentView.setNeedsLayout()
-            reusableCell.contentView.layoutIfNeeded()
-        }
-        return cell
-    }
-    
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        1
+        tableView.delegate = tableAdapter
+        tableView.dataSource = tableAdapter
+        tableView.rowHeight = UITableView.automaticDimension
+        tableView.estimatedRowHeight = 50
     }
 }
 
@@ -65,7 +37,7 @@ extension TaskDetailScreenViewController: TaskDetailScreenViewInput {
     func handleState(_ state: TaskDetailScreenModels.State) {
         switch state {
         case .success(let currentTask):
-            self.currentTask = currentTask
+            tableAdapter.currentTask = currentTask
             tableView.reloadData()
         default:
             return
